% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse.edf.signal.header}
\alias{parse.edf.signal.header}
\title{Parse the signal header of a European Data Format (EDF and EDF+) file.}
\usage{
parse.edf.signal.header(data, header.global)
}
\arguments{
\item{data}{The n*256 bytes of raw data that contain the EDF/EDF+ signal header,}

\item{header.global}{The global EDF/EDF+ header.
#' where n is the number of signals in the EDF/EDF+ file.}
}
\value{
A named list containing the information in the EDF/EDF+ signal header.
}
\description{
This function parses the signal header of an EDF or EDF+ file.
}
\keyword{internal}
