% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse.event}
\alias{parse.event}
\title{Parse one EDF+ Time-stamped Annotation List (TAL).}
\usage{
parse.event(data)
}
\arguments{
\item{data}{The bytes of raw data that contain the TAL.}
}
\value{
A data frame containing the information in the TAL: onset, duration and annotation.
}
\description{
This function parses one EDF+ Time-stamped Annotation List (TAL).
}
\keyword{internal}
