% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edf.char.to.num}
\alias{edf.char.to.num}
\title{Convert string to a numeric.}
\usage{
edf.char.to.num(s)
}
\arguments{
\item{s}{The string to be converted}
}
\value{
A numeric.
}
\description{
This function removes spaces from a string and converts it to a numeric.
This is required when reading EDF data, since the data fields in the header
are of fixed size, and extra whitespace must hence be trimmed.
}
\examples{
s <- "123   "
edf:::edf.char.to.num(s)
}
\keyword{internal}

