% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse.edf.annotations}
\alias{parse.edf.annotations}
\title{Parse the annotation data in an EDF+ file.}
\usage{
parse.edf.annotations(data)
}
\arguments{
\item{data}{The data containining the annotations (as 8-bit unsigned integers).}
}
\value{
A data frame containing the annotatations in the EDF+ file.
}
\description{
This function parses the annotations in an EDF+ file.
}
\keyword{internal}

