% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse.edf.global.header}
\alias{parse.edf.global.header}
\title{#' Parse the global header of a European Data Format (EDF and EDF+) file.}
\usage{
parse.edf.global.header(data)
}
\arguments{
\item{data}{The 256 bytes of raw data that contain the EDF/EDF+ header.}
}
\value{
A named list containing the information in the EDF/EDF+ header.
}
\description{
This function parses the header of an EDF or EDF+ file.
}
\keyword{internal}

