% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim.end}
\alias{trim.end}
\title{Remove trailing whitespace from a string.}
\usage{
trim.end(s)
}
\arguments{
\item{s}{The string to be converted}
}
\value{
The string gives as input without trailing whitespace
}
\description{
This function removes trailing whitespace from a string.
This is required when reading EDF data, since the data fields in the header
are of fixed size, and extra whitespace must hence be trimmed.
}
\examples{
s <- "abc   "
edf:::trim.end(s)
}
\keyword{internal}

